pyproject.toml
[tool.poetry]
name = "dtc"
version = "0.1.0"
description = ""
authors = ["Tristan Bester <tristanbester@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.4.3"
jupyter = "^1.0.0"
notebook = "^6.4.12"
numpy = "^1.23.1"
matplotlib = "^3.5.2"
python-dotenv = "^0.20.0"
wandb = "^0.13.1"

[tool.poetry.dev-dependencies]
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
